--- ips_4_1.txt
@@ -1,4 +1,0 @@
-127.0.0.1:9000
-127.0.0.1:9001
-127.0.0.1:9002
-127.0.0.1:9003
