--- ips_8_3.txt
@@ -1,8 +1,0 @@
-127.0.0.1:9000
-127.0.0.1:9001
-127.0.0.1:9002
-127.0.0.1:9003
-127.0.0.1:9004
-127.0.0.1:9005
-127.0.0.1:9006
-127.0.0.1:9007
